+++ app/components/data-menu/services/data-service.js
@@ -276,6 +276,9 @@ angular.module('data-menu')
        */
       var getTimeSeriesForObject = function (objectId, start, end, defer) {
 
+        // maximum number of timeseries events, more probably results in a 
+        // memory error.
+        var MAX_NR_TIMESERIES_EVENTS = 25000;  
         var promise = TimeseriesService.getTimeseries(objectId, {
           start: start,
           end: end
@@ -287,7 +290,8 @@ angular.module('data-menu')
             // the timeseries with too little measurements...
             var filteredResult = [];
             angular.forEach(result, function (value) {
-              if (value.events.length > 1) {
+              if (value.events.length > 1 && 
+                  value.events.length < MAX_NR_TIMESERIES_EVENTS) {
                 filteredResult.push(value);
               }
             });